New Debug Command, Inspiration Obj Scaffolding
Added a new debugging command so I can quickly skip the intro-- I may edit it after full release so ANYONE can use it. Seems like a nice QOL thing.

Adding the scaffolding for inspiration objects so that when one is consumed at the right time (when all have ediblebit) it will remove the others, and  have the emotional changes affect the player.