Inspo rendering setup, dreamlands multipaths setup
Working on the dreamlands some. I need to be careful and remember not to creep too much content. Start small.

Started the infrastructure so that the dreamlands will randomly pick paths north, south and west of the somber planes. Created the blood dam, and I think it will be the first inspo for the prologue.