Changed CALL-OF-CARCOSA? routines reformed, talking SHOULD work as intended
Testing needed ofc. Changed routines like CALL-OF-CARCOSA? et all which will check for influence on a dream, and will save result into a global variable. This is mostly for convenience because TONS of places are gonna have trigger routines that look at the influence. Having them be globals makes sense. Talking person object tempalte SHOULD work now as intended. Uses a property TALKTBL (which ought to be an LTABLE of choice rooms). Will sequentially go from first elem to last, checking if it 1) hasn't been visited before 2) its trigger routine is true and is both ARE true, we'll use that convo node for the choice-room based conversation.